requirements.txt
Homie4
paho-mqtt~=1.6.1
pytest